refactor: move RacingManager to domain package
